Strip HTML-part from multipart/alternative before decrypting inline.
The ciphertext inside of that mime-part is probably wrapped in HTML,
which GnuPG chokes upon, so we would have to parse the HTML to handle
the message-part properly.
Also it's not clear how to handle the resulting plain-text: is it HTML
or simple text?  That depends on the sending MUA and the original
input.
In summary, that's too much complications.